requirements.txt
atlassian-python-api
requests
openai